Refactor into ES modules
