+++ index.html
@@ -721,7 +721,7 @@
       </div>
     </div>
 
-    <script src="src/oblix.js"></script>
+    <script type="module" src="src/main.js"></script>
     <canvas
       id="watermarkCanvas"
       width="1"
